Refactored components + added absolute path instead of relative for imports
