+++ src/components/GameViews/GameFinishedView.vue
@@ -16,17 +16,17 @@
 </template>
 
 <script>
-import PlayerDisplay from "./PlayerDisplay";
+import PlayerDisplay from "@/components/GameComponents/PlayerDisplay";
 
 export default {
   data: () => ({}),
   props: {
     room_data: {
-      required: true
-    }
+      required: true,
+    },
   },
   components: {
-    "player-display": PlayerDisplay
+    "player-display": PlayerDisplay,
   },
   methods: {},
   computed: {
@@ -36,7 +36,7 @@ export default {
         return a.points < b.points ? 1 : -1;
       });
       return players;
-    }
-  }
+    },
+  },
 };
 </script>
