+++ src/components/GameViews/SessionDisplay.vue
