+++ vue-config.js
@@ -1,0 +1,10 @@
+var path = require('path');
+module.exports = {
+  configureWebpack: {
+    resolve: {
+      alias: {
+        src: path.resolve(__dirname, 'src'),
+      },
+    },
+  },
+};
